Delete the central ClearDatabase function. Now clear DBs by individual clear() on each model

+++ models/checklist.js
@@ -1,4 +1,3 @@
-//const { dbChecklists } = require('../database/createDB')
 const mongoose = require('mongoose')
 
 const checklistSchema = new mongoose.Schema({
@@ -15,41 +14,38 @@ const checklistSchema = new mongoose.Schema({
 
 const ChecklistModel = mongoose.model('Checklist', checklistSchema)
 
-async function clear() {
-  return await ChecklistModel.deleteMany({})
-}
-
 
 /** UserModel functions **/
 
-// Mongo ok
 async function findChecklists(query) {
   return await ChecklistModel.find(query)
 }
 
-// Mongo ok
 async function saveChecklist(checklist) {
   const newChecklist = new ChecklistModel(checklist)
   //console.log('newChecklist', newChecklist)
   return await newChecklist.save()
 }
 
-// Mongo
+
 async function findChecklist(query) {
   const checklist = await ChecklistModel.findOne(query)
   // console.log('checklist i model', checklist)
   return checklist
 }
 
-// Mongo
 async function removeChecklist(query) {
   const response = await ChecklistModel.deleteMany(query)
   return response.deletedCount
 }
 
-// Mongo ok
 async function updateChecklist(query, toUpdate) {
   return await ChecklistModel.findOneAndUpdate(query, toUpdate, { new: true })
+}
+
+// Only clear on tests
+async function clear() {
+  if (process.env.ENVIRONMENT === 'test') return await ChecklistModel.deleteMany({})
 }
 
 module.exports = {
